refactor: consolidate misc_info stream with conditional CPU frequency
- Extract misc_info logic to common/streams/misc_info.rs
- Move shared structs: TimeValue, MachTaskBasicInfo, TaskThreadsTimeInfo
- Create TaskDumperHelper trait for platform-specific task operations
- Apply #[cfg(target_os = "macos")] for CPU frequency code
- Both iOS and macOS now use the common implementation
- Reduces ~180 lines of duplicate code

Key differences handled:
- macOS: Full CPU frequency support via sysctl
- iOS: CPU frequency fields remain 0 (system restrictions)

This demonstrates conditional compilation pattern for platform differences.
