+++ src/apple/common/streams/mod.rs
@@ -1,1 +1,2 @@
 pub mod breakpad_info;
+pub mod misc_info;
