+++ src/apple/ios/streams/misc_info.rs
@@ -1,76 +1,20 @@
 use crate::{
     apple::{
-        common::mach,
+        common::streams::misc_info::{self, TaskDumperHelper},
         ios::{minidump_writer::MinidumpWriter, task_dumper::TaskDumper},
     },
     dir_section::DumpBuf,
-    mem_writer::*,
-    minidump_format::{
-        format::{MiscInfoFlags, MINIDUMP_MISC_INFO_2 as MDRawMiscInfo},
-        MDRawDirectory, MDStreamType,
-    },
+    minidump_format::MDRawDirectory,
 };
-use std::time::Duration;
 
-/// From <usr/include/mach/time_value.h>
-#[repr(C)]
-#[derive(Copy, Clone)]
-struct TimeValue {
-    seconds: i32,
-    microseconds: i32,
-}
-
-impl From<TimeValue> for Duration {
-    fn from(tv: TimeValue) -> Self {
-        let mut seconds = tv.seconds as u64;
-        let mut microseconds = tv.microseconds as u32;
-        // This _probably_ will never happen, but this will avoid a panic in
-        // Duration::new() if it does
-        if tv.microseconds >= 1000000 {
-            seconds += 1;
-            microseconds -= 1000000;
-        }
-
-        Duration::new(seconds, microseconds * 1000)
+impl TaskDumperHelper for TaskDumper {
+    fn pid_for_task(&self) -> Result<i32, String> {
+        self.pid_for_task().map_err(|e| e.to_string())
     }
-}
 
-/// From <usr/include/mach/task_info.h>, this includes basic information about
-/// a task.
-#[repr(C, packed(4))]
-struct MachTaskBasicInfo {
-    /// Virtual memory size in bytes
-    virtual_size: u64,
-    /// Resident memory size in bytes
-    resident_size: u64,
-    /// Maximum resident memory size in bytes
-    resident_size_max: u64,
-    /// Total user run time for terminated threads
-    user_time: TimeValue,
-    /// Total system run time for terminated threads
-    system_time: TimeValue,
-    /// Default policy for new threads
-    policy: i32,
-    /// Suspend count for task
-    suspend_count: i32,
-}
-
-impl mach::TaskInfo for MachTaskBasicInfo {
-    const FLAVOR: u32 = mach::task_info::MACH_TASK_BASIC_INFO;
-}
-
-/// From <usr/include/mach/task_info.h>, this includes times for currently
-/// live threads in the task.
-#[repr(C, packed(4))]
-struct TaskThreadsTimeInfo {
-    /// Total user run time for live threads
-    user_time: TimeValue,
-    /// total system run time for live threads
-    system_time: TimeValue,
-}
-
-impl mach::TaskInfo for TaskThreadsTimeInfo {
-    const FLAVOR: u32 = mach::task_info::TASK_THREAD_TIMES_INFO;
+    fn task_info<T: crate::apple::common::mach::TaskInfo>(&self) -> Result<T, String> {
+        self.task_info().map_err(|e| e.to_string())
+    }
 }
 
 impl MinidumpWriter {
@@ -85,83 +29,7 @@ impl MinidumpWriter {
         buffer: &mut DumpBuf,
         dumper: &TaskDumper,
     ) -> Result<MDRawDirectory, super::super::WriterError> {
-        let mut info_section = MemoryWriter::<MDRawMiscInfo>::alloc(buffer)
-            .map_err(|e| super::super::WriterError::MemoryWriterError(e.to_string()))?;
-        let dirent = MDRawDirectory {
-            stream_type: MDStreamType::MiscInfoStream as u32,
-            location: info_section.location(),
-        };
-
-        let pid = dumper
-            .pid_for_task()
-            .map_err(|e| super::super::WriterError::TaskDumperError(e.to_string()))?;
-
-        let mut misc_info = MDRawMiscInfo {
-            size_of_info: std::mem::size_of::<MDRawMiscInfo>() as u32,
-            flags1: MiscInfoFlags::MINIDUMP_MISC1_PROCESS_ID.bits()
-                | MiscInfoFlags::MINIDUMP_MISC1_PROCESS_TIMES.bits(),
-            process_id: pid as u32,
-            process_create_time: 0,
-            process_user_time: 0,
-            process_kernel_time: 0,
-            processor_max_mhz: 0,
-            processor_current_mhz: 0,
-            processor_mhz_limit: 0,
-            processor_max_idle_state: 0,
-            processor_current_idle_state: 0,
-        };
-
-        // Get process start time
-        // SAFETY: syscall
-        misc_info.process_create_time = unsafe {
-            let mut proc_info = std::mem::MaybeUninit::<libc::proc_bsdinfo>::uninit();
-            let size = std::mem::size_of::<libc::proc_bsdinfo>() as i32;
-            if libc::proc_pidinfo(
-                pid,
-                libc::PROC_PIDTBSDINFO,
-                0,
-                proc_info.as_mut_ptr().cast(),
-                size,
-            ) == size
-            {
-                let proc_info = proc_info.assume_init();
-                proc_info.pbi_start_tvsec as u32
-            } else {
-                0
-            }
-        };
-
-        // Get task timing information
-        let basic_info = dumper.task_info::<MachTaskBasicInfo>().ok();
-        let thread_times_info = dumper.task_info::<TaskThreadsTimeInfo>().ok();
-
-        let user_time = basic_info
-            .as_ref()
-            .map(|bi| Duration::from(bi.user_time))
-            .unwrap_or_default()
-            + thread_times_info
-                .as_ref()
-                .map(|tt| Duration::from(tt.user_time))
-                .unwrap_or_default();
-        let system_time = basic_info
-            .as_ref()
-            .map(|bi| Duration::from(bi.system_time))
-            .unwrap_or_default()
-            + thread_times_info
-                .as_ref()
-                .map(|tt| Duration::from(tt.system_time))
-                .unwrap_or_default();
-
-        misc_info.process_user_time = user_time.as_secs() as u32;
-        misc_info.process_kernel_time = system_time.as_secs() as u32;
-
-        // Note: CPU frequency information is typically not available on iOS
-        // due to system restrictions, so we leave those fields as 0
-
-        info_section
-            .set_value(buffer, misc_info)
-            .map_err(|e| super::super::WriterError::MemoryWriterError(e.to_string()))?;
-
-        Ok(dirent)
+        misc_info::write_misc_info(buffer, dumper)
+            .map_err(|e| super::super::WriterError::MemoryWriterError(e.to_string()))
     }
 }
